별거 아닌 수정 /api/login 으로 하면 일반 로그인이 안되서 /login으로 변경 initDb와 Cart 데이터를 담을 DTO 생성중

+++ src/main/java/com/example/project/dto/store/StoreListViewResponse.java
@@ -12,7 +12,7 @@ public class StoreListViewResponse {
     private final String boss;
     private final String picture;
     private final String content;
-    private final float rating;
+    private final double rating;
 
     public StoreListViewResponse(Store store) {
         this.id = store.getId();
